Returns the value of x raised to the power of y

+++ 0x08-recursion/4-pow_recursion.c
@@ -17,5 +17,5 @@ int _pow_recursion(int x, int y)
 	{
 		return (-1);
 	}
-	return (pow(x, y));
+	return (powf(x, y));
 }
